rename: PreferenceManager 패키지 수정
