+++ app/src/main/java/com/devik/homebarorder/data/source/local/database/PreferenceManager.kt
@@ -1,4 +1,4 @@
-package com.devik.homebarorder.data.source.local
+package com.devik.homebarorder.data.source.local.database
 
 import android.content.Context
 import com.devik.homebarorder.util.Constants.PREFERENCE_NAME
